backend/requirements.txt
-r ../requirements.txt
alembic==1.16.5
fastapi==0.116.1
PyJWT==2.10.1
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
